+++ components/Register/LeftDisplay.tsx
