+++ components/shared/EventCard.tsx
