+++ pages/volunteerHistory.tsx
@@ -1,4 +1,4 @@
-import LongEventCard from '@/components/LongEventCard';
+import LongEventCard from '@/components/shared/LongEventCard';
 import React from 'react';
 import Image from 'next/image';
 import useSWR from 'swr';
